refactor: extract ComboBoxBase from GeneratedVaadinComboBox
